@@ -3,24 +3,21 @@ name = "foxel"
 version = "1"
 description = "foxel.cc"
 readme = "README.md"
-requires-python = ">=3.13"
+requires-python = ">=3.14"
 dependencies = [
-    "aioboto3>=15.2.0",
-    "aiofiles>=25.1.0",
-    "fastapi>=0.116.1",
-    "passlib[bcrypt]>=1.7.4",
-    "bcrypt>=3.2.2,<6.0",
-    "pillow>=11.3.0",
-    "pyjwt>=2.10.1",
-    "pysocks>=1.7.1",
-    "python-dotenv>=1.1.1",
-    "python-multipart>=0.0.20",
-    "qdrant-client>=1.15.1",
-    "rawpy>=0.25.1",
-    "telethon>=1.41.2",
-    "tortoise-orm>=0.25.2",
-    "uvicorn>=0.37.0",
-    "pymilvus[milvus-lite]>=2.6.2",
+    "aioboto3>=15.5.0",
+    "bcrypt>=5.0.0",
+    "fastapi>=0.127.0",
     "paramiko>=4.0.0",
-    "pydantic[email]>=2.11.7",
+    "pillow>=12.0.0",
+    "pydantic[email]>=2.12.5",
+    "pyjwt>=2.10.1",
+    "pymilvus[milvus-lite]>=2.6.5",
+    "pysocks>=1.7.1",
+    "python-dotenv>=1.2.1",
+    "python-multipart>=0.0.21",
+    "qdrant-client>=1.16.2",
+    "telethon>=1.42.0",
+    "tortoise-orm>=0.25.3",
+    "uvicorn>=0.40.0",
 ]
